put test logic to a particular method
